Refactoring logical operators to include support for $not
Side effect is that $not is used whenever a comparison uses the '!' in Java.
This means that for an operation such as '!someField.equals("foo")', the result
will include both documents having a field "someField' with a value different of "foo",
but also, the result will include all documents NOT having the aformentionned 'someField' field.